+++ src/domain/project-manifest.ts
@@ -1,4 +1,5 @@
 import path from "node:path";
+import { z } from "zod";
 import { removeRecordKey } from "../utils/record-utils";
 import { removeTrailingSlash } from "../utils/string-utils";
 import { DomainName } from "./domain-name";
@@ -222,4 +223,18 @@ export function removeTestable(
  */
 export function manifestPathFor(projectPath: string): string {
   return path.join(projectPath, "Packages/manifest.json");
+}
+
+// TODO: Add a better schema
+const projectManifestSchema = z.object({}).passthrough();
+
+/**
+ * Parses the content of a `manifest.json` file to a {@link UnityProjectManifest}.
+ * @param content The files content.
+ * @returns The parsed file.
+ * @throws {Error} If parsing failed.
+ */
+export function parseProjectManifest(content: string): UnityProjectManifest {
+  const json = JSON.parse(content);
+  return projectManifestSchema.parse(json) as UnityProjectManifest;
 }
